backend/requirements.txt
Django==3.2.13
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
psycopg2==2.9.3
python-dotenv==0.19.2
requests==2.27.1